steps: Move the shared behave steps to a new shared location
Moves the generic behave steps that will certainly be needed in other
test suites than dnf to the new `common` module.

For the output-checking functions I've moved the exact matches and
line-by-line regex checks. Ideally those should be sufficient. More can
be moved later if (absolutely :)) necessary.

Includes a couple minor changes that are needed to make it work.

The modules that contain behave steps need to be listed in
common/__init__.py so that they are explicitly loaded by python and the
behave steps in them are actually added to behave. And then it is better
to explicitly import common in environment.py to make sure the module is
loaded altogether (insted of relying on a different import to load it).
