+++ dnf-behave-tests/common/__init__.py
@@ -1,0 +1,4 @@
+# make behave load steps from these modules
+import common.cmd
+import common.file
+import common.output
